requirements.txt
fastapi[standard]==0.115.12
pydantic==2.11.4
pillow==11.2.1